improve architecture of files:
remove duplicata and too generic files.
Instead move helper function to dedicated files.
