+++ libvirt/network_def_test.go
@@ -170,3 +170,29 @@ func TestNetworkFromLibvirt(t *testing.T) {
 		t.Errorf("Wrong forward mode: %s", dn.Forward.Mode)
 	}
 }
+
+func TestGetHostXMLDesc(t *testing.T) {
+	ip := "127.0.0.1"
+	mac := "XX:YY:ZZ"
+	name := "localhost"
+
+	data := getHostXMLDesc(ip, mac, name)
+
+	dd := libvirtxml.NetworkDHCPHost{}
+	err := xml.Unmarshal([]byte(data), &dd)
+	if err != nil {
+		t.Errorf("error %v", err)
+	}
+
+	if dd.IP != ip {
+		t.Errorf("expected ip %s, got %s", ip, dd.IP)
+	}
+
+	if dd.MAC != mac {
+		t.Errorf("expected mac %s, got %s", mac, dd.MAC)
+	}
+
+	if dd.Name != name {
+		t.Errorf("expected name %s, got %s", name, dd.Name)
+	}
+}
