--- libvirt/utils_libvirt.go
@@ -1,45 +1,0 @@
-package libvirt
-
-import (
-	"encoding/xml"
-	"log"
-
-	libvirt "github.com/libvirt/libvirt-go"
-	libvirtxml "github.com/libvirt/libvirt-go-xml"
-)
-
-func getHostArchitecture(virConn *libvirt.Connect) (string, error) {
-	type HostCapabilities struct {
-		XMLName xml.Name `xml:"capabilities"`
-		Host    struct {
-			XMLName xml.Name `xml:"host"`
-			CPU     struct {
-				XMLName xml.Name `xml:"cpu"`
-				Arch    string   `xml:"arch"`
-			}
-		}
-	}
-
-	info, err := virConn.GetCapabilities()
-	if err != nil {
-		return "", err
-	}
-
-	capabilities := HostCapabilities{}
-	xml.Unmarshal([]byte(info), &capabilities)
-
-	return capabilities.Host.CPU.Arch, nil
-}
-
-func getHostCapabilities(virConn *libvirt.Connect) (libvirtxml.Caps, error) {
-	// We should perhaps think of storing this on the connect object
-	// on first call to avoid the back and forth
-	caps := libvirtxml.Caps{}
-	capsXML, err := virConn.GetCapabilities()
-	if err != nil {
-		return caps, err
-	}
-	xml.Unmarshal([]byte(capsXML), &caps)
-	log.Printf("[TRACE] Capabilities of host \n %+v", caps)
-	return caps, nil
-}
